lj_crecord.c: JIT pointer difference for all element sizes
Support JIT for taking the difference between two pointers whose
element sizes are not a power of 2. This was previously an NYI but not
generates code using a division to convert a pointer to an element
number.

The case where the element size is a power of 2 is handled the same
way as before i.e. using a bit-shift instead of a division.

(cherry picked from commit cf655d95a5ab65e8590f1aae24dfa0cdf0c9299b)

# Conflicts:
#	src/lj_crecord.c