lj_record.c: Record IFUNC/IFUNCV the same as FUNC/FUNCV
Record calls to blacklisted functions the same way as calls to
uncompiled functions. This allows for the possibility that a
blacklisted function can be successfully recorded as part of a larger
trace.

If the function truly cannot be compiled into other traces then it
will continue to abort and lead to more blacklistings, at the cost of
some extra "due diligence" by the JIT to avoid false-negatives. This
is perfectly in line with the RaptorJIT goal of blacklisting code
conservatively and taking pains to find successful traces.

This change does not affect the treatment blacklisted loops. Those may
or may not deserve further consideration.

Resolves raptorjit/raptorjit#119.

(cherry picked from commit f5d810e97c1d38f3ead28f965c328dc7eac487e9)